requirements.txt
streamlit
psycopg2-binary
numpy
pandas
sentence-transformers
pgvector
python-dotenv
datasets
